Move tasks to package so that they can be imported
