+++ src/main/java/HdrToCsv.java
@@ -1,10 +1,8 @@
-import org.HdrHistogram.Histogram;
 import org.kohsuke.args4j.Option;
-import psy.lob.saw.OrderedHistogramLogReader;
+import psy.lob.saw.CsvConverter;
 
 import java.io.*;
 import java.nio.file.Paths;
-import java.util.Locale;
 
 public class HdrToCsv implements Runnable
 {
@@ -35,45 +33,11 @@ public class HdrToCsv implements Runnable
     {
         try(InputStream inputStream = new FileInputStream(inputFile))
         {
-            new Converter(inputStream).convert();
+            new CsvConverter(inputStream).convert();
         }
         catch (IOException e)
         {
             throw new RuntimeException(e);
-        }
-    }
-
-    public static class Converter
-    {
-        private final InputStream inputStream;
-
-        public Converter(InputStream inputStream)
-        {
-            this.inputStream = inputStream;
-        }
-
-        public void convert() throws FileNotFoundException
-        {
-            OrderedHistogramLogReader reader = new OrderedHistogramLogReader(inputStream);
-            System.out.println(
-                    "#Absolute timestamp,Relative timestamp,Throughput,Min,Avg,p50,p90,p95,p99,p999,p9999,Max");
-            while (reader.hasNext())
-            {
-                Histogram interval = (Histogram) reader.nextIntervalHistogram();
-                System.out.printf(Locale.US,
-                        "%.3f,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d%n",
-                        interval.getStartTimeStamp() / 1000.0,
-                        interval.getStartTimeStamp() / 1000 - (long) reader.getStartTimeSec(),
-                        interval.getTotalCount(), interval.getMinValue(),
-                        (long) interval.getMean(),
-                        interval.getValueAtPercentile(50),
-                        interval.getValueAtPercentile(90),
-                        interval.getValueAtPercentile(95),
-                        interval.getValueAtPercentile(99),
-                        interval.getValueAtPercentile(99.9),
-                        interval.getValueAtPercentile(99.99),
-                        interval.getMaxValue());
-            }
         }
     }
 }
